Updated workflow (#305)
Migrated to Github Actions and bumped the tooling versions as and where needed. 
Made the respective code changed to satiate static analysis tooling

+++ classes/helpers/NostoHelperCurrency.php
@@ -103,6 +103,7 @@ class NostoHelperCurrency
         $allCurrencies = Currency::getCurrenciesByIdShop(NostoHelperContext::getShopId());
         if ($onlyActive === true) {
             $currencies = array();
+            // @phan-suppress-next-line PhanTypeSuspiciousNonTraversableForeach
             foreach ($allCurrencies as $currency) {
                 if (self::currencyActive($currency)) {
                     $currencies[] = $currency;
@@ -239,7 +240,7 @@ class NostoHelperCurrency
             $currencyDecimalsEnabled = $currencyObject->decimals;
         }
 
-        return $currencyDecimalsEnabled * _PS_PRICE_DISPLAY_PRECISION_;
+        return $currencyDecimalsEnabled * (int) _PS_PRICE_DISPLAY_PRECISION_;
     }
 
     private static function currencyActive(array $currency)
